refactor(arch): decompose ServerDaemon.cpp into daemon/ folder by responsibility
- Split monolithic 646-line ServerDaemon.cpp into 4 focused modules under daemon/:
  * daemon.cpp: constructor and destructor (~40 lines)
  * lifecycle.cpp: start(), stop(), isRunning(), shutdownRequested() (~140 lines)
  * protocol.cpp: acceptLoop(), handleClient(), processRequest() (~400 lines)
  * helpers.cpp: trim(), asBytes(), sendAll(), serverDaemonError() (~50 lines)

Structure improvements:
- No name redund dance: all files live under daemon/ subfolder
- Each file <= 200 lines, well under policy threshold ~200
- Public header (ServerDaemon.hpp) remains unchanged
- CMakeLists.txt updated to link all daemon/ modules

Zero functional change - identical binary behavior, improved maintainability.
