+++ .gitignore
@@ -29,6 +29,11 @@ Thumbs.db
 # Tooling cache
 __pycache__/
 *.pyc
+.venv/
+.mypy_cache/
+.ruff_cache/
+.coverage
+.hypothesis/
 .pytest_cache/
 .pytest-basetemp*/
 .pytest_tmp*/
